Array_Sum.py
You are given an array A consisting of N integers. 

Task

Print the sum of the elements in the array. 

Note: Some of the integers may be quite large.

Input Format

The first line contains a single integer N denoting the size of the array.  
The next line contains space-separated integers denoting the elements of the array.
Output format

Print a single value representing the sum of the elements in the array.